requirements.txt
# Core dependencies
langchain_openai
langchain-core
langchain
python-dotenv
pydantic

# Graph-based agent framework
langgraph

# Model dependencies
transformers
accelerate
bitsandbytes